Full Support for Variable & Constant Declarations
^ Incremented support for variable declarations.
All variable declarations are now early bound in the program. This means that both global and local variables will have lexical scoping checks during compilation, and not during runtime.

This also means that the `globals` HashMap that was used in the VM before no longer exists. Instead, all variable values are saved on the stack, and are referred by their index on the stack.

+ Added support full support for constant declarations.

+ Added support for blocks and block scoping.